ncjoin for Python
ROMS outputs are done as a collection subdomains of the larger grid, therefore requiring them to be compiled together for the full netcdf. Users of ROMS may be familiar with ncjoin, a script designed to take individual .nc files and join them into their summed image - this is the python analog.

This tool is mainly for those wanting to use ROMS outputs, particularly through Python, but have not yet joined subdomains.